change PaymentRequest model and refactor api
