+++ prisma/schema.prisma
@@ -116,19 +116,20 @@ model Notification {
 }
 
 model Token {
-  id              String           @id
-  token           String
-  createdAt       DateTime         @default(now())
-  redeemedAt      DateTime?
-  gift            String?
-  Notification    Notification[]
-  createdByPubkey String?
-  recipientPubkey String?
+  id               String          @id
+  token            String
+  createdAt        DateTime        @default(now())
+  redeemedAt       DateTime?
+  gift             String?
+  Notification     Notification[]
+  createdByPubkey  String?
+  recipientPubkey  String?
   // whether or not this token is a fee
-  isFee           Boolean          @default(false)
-  recipient       User?            @relation(fields: [recipientPubkey], references: [pubkey], name: "TokenRecipient")
-  createdBy       User?            @relation(fields: [createdByPubkey], references: [pubkey], name: "TokenCreatedBy")
-  PaymentRequest  PaymentRequest[]
+  isFee            Boolean         @default(false)
+  recipient        User?           @relation(fields: [recipientPubkey], references: [pubkey], name: "TokenRecipient")
+  createdBy        User?           @relation(fields: [createdByPubkey], references: [pubkey], name: "TokenCreatedBy")
+  paymentRequest   PaymentRequest? @relation(fields: [paymentRequestId], references: [id])
+  paymentRequestId String?
 }
 
 model MintlessTransaction {
@@ -210,13 +211,13 @@ model PaymentRequest {
   id         String   @id @default(uuid())
   amount     Int?
   userPubkey String
-  txid       String?
   paid       Boolean  @default(false)
   reusable   Boolean
   createdAt  DateTime @default(now())
   updatedAt  DateTime @updatedAt
   user       User     @relation(fields: [userPubkey], references: [pubkey])
-  token      Token?   @relation(fields: [txid], references: [id])
+  // multiple tokens if the payment request is reusable
+  tokens     Token[]
 }
 
 // model GiftCondition {
